refactor(backend): move notes/media endpoints into farm/notes
Move MediaFileUploadView, NoteAttachmentListCreateView,
NoteAttachmentDeleteView and the media upload constants from
farm/views.py, and NoteAttachmentSerializer from farm/serializers.py,
into a farm/notes domain package. Route table unchanged; test mock
patch targets repointed to farm.notes.views.
