+++ backend/farm/notes/__init__.py
@@ -1,0 +1,1 @@
+"""Notes and media attachments domain: uploads and note image attachments."""
