+++ backend/farm/serializers.py
@@ -33,7 +33,6 @@ from .models import (
     Field,
     Location,
     MediaFile,
-    NoteAttachment,
     PlantingPlan,
     Supplier,
     Task,
@@ -1452,62 +1451,6 @@ class SeedDemandSerializer(serializers.Serializer):
     package_suggestion = SeedDemandPackageSuggestionSerializer(allow_null=True, required=False)
     packages_needed = serializers.IntegerField(allow_null=True, required=False)
     warning = serializers.CharField(allow_null=True)
-
-
-class NoteAttachmentSerializer(serializers.ModelSerializer):
-    """Serializer for note image attachments."""
-
-    image_url = serializers.SerializerMethodField()
-    created_by_user = AuditUserSerializer(source='created_by', read_only=True)
-    updated_by_user = AuditUserSerializer(source='updated_by', read_only=True)
-
-    def get_image_file(self, obj):
-        if not obj.image_file_id:
-            return None
-        return {
-            'id': obj.image_file_id,
-            'storage_path': obj.image_file.storage_path,
-        }
-
-    class Meta:
-        model = NoteAttachment
-        fields = [
-            "id",
-            "planting_plan",
-            "image",
-            "image_url",
-            "caption",
-            "created_at",
-            "updated_at",
-            "created_by",
-            "updated_by",
-            "created_by_user",
-            "updated_by_user",
-            "width",
-            "height",
-            "size_bytes",
-            "mime_type",
-        ]
-        read_only_fields = [
-            "id",
-            "planting_plan",
-            "created_at",
-            "updated_at",
-            "created_by",
-            "updated_by",
-            "width",
-            "height",
-            "size_bytes",
-            "mime_type",
-        ]
-
-    def get_image_url(self, obj):
-        request = self.context.get("request")
-        if not obj.image:
-            return None
-        if request:
-            return request.build_absolute_uri(obj.image.url)
-        return obj.image.url
 
 
 class CultureHistoryEntrySerializer(serializers.Serializer):
