+++ backend/farm/views.py
@@ -18,19 +18,18 @@ from django.db.models import Case, When, Value, F, FloatField, IntegerField, Exp
 from django.db.models.functions import Coalesce, Ceil, Cast
 from django.http import HttpResponseBadRequest, HttpResponseForbidden
 from django.shortcuts import get_object_or_404, redirect
-from rest_framework import viewsets, status, generics, parsers, permissions
+from rest_framework import viewsets, status, generics, permissions
 from rest_framework.decorators import action
 from rest_framework.exceptions import ValidationError as DRFValidationError
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from django.core.files.storage import default_storage
 from django.utils import timezone, translation
 from django.utils.crypto import get_random_string
 from django.utils.text import slugify
 from django.utils.dateparse import parse_date
 from django.utils.translation import gettext as _
-from .models import AgentLoginToken, Location, Field, Bed, BedLayout, FieldLayout, Culture, CultureSupplierData, PlantingPlan, Task, Supplier, NoteAttachment, MediaFile, SeedPackage, PublicCulture, culture_media_upload_path, EntityRevision, Project, ProjectMembership, ProjectInvitation
+from .models import AgentLoginToken, Location, Field, Bed, BedLayout, FieldLayout, Culture, CultureSupplierData, PlantingPlan, Task, Supplier, MediaFile, SeedPackage, PublicCulture, EntityRevision, Project, ProjectMembership, ProjectInvitation
 from .common.mixins import ProjectRevisionMixin, ProjectScopedMixin
 from .history import (
     _build_entity_revision_changes,
@@ -53,7 +52,6 @@ from .serializers import (
     SupplierSerializer,
     SeedDemandSerializer,
     PublicCultureSerializer,
-    NoteAttachmentSerializer,
     CultureHistoryEntrySerializer,
     CultureRestoreSerializer,
     SeedPackageSerializer,
@@ -83,11 +81,6 @@ from .services.project_invitations import (
 
 from .services_area import calculate_remaining_bed_area
 
-from .image_processing import (
-    process_note_image,
-    ImageProcessingError,
-    ImageProcessingBackendUnavailableError,
-)
 from .services.seed_demand import build_seed_demand_rows, parse_selected_suppliers
 from .services.public_cultures import (
     DuplicatePublicCultureError,
@@ -100,15 +93,6 @@ from config.frontend_urls import build_public_frontend_url
 
 
 logger = logging.getLogger(__name__)
-
-MAX_MEDIA_UPLOAD_BYTES = 10 * 1024 * 1024
-ALLOWED_MEDIA_UPLOAD_CONTENT_TYPES = {
-    'image/jpeg',
-    'image/png',
-    'image/webp',
-    'image/gif',
-}
-
 
 def _invitation_error_response(exc: InvitationFlowError) -> Response:
     """Build a consistent error response for invitation domain errors."""
@@ -1818,114 +1802,6 @@ class GlobalHistoryRestoreView(APIView):
             culture.save()
 
         return Response(CultureSerializer(culture).data, status=status.HTTP_200_OK)
-
-
-class MediaFileUploadView(APIView):
-    parser_classes = [parsers.MultiPartParser, parsers.FormParser]
-
-    def post(self, request):
-        active_project = get_active_project_or_400(request)
-        upload = request.FILES.get('file')
-        if upload is None:
-            return Response({'file': ['This field is required.']}, status=status.HTTP_400_BAD_REQUEST)
-        if upload.size > MAX_MEDIA_UPLOAD_BYTES:
-            return Response({'file': ['File is too large. Maximum allowed size is 10 MB.']}, status=status.HTTP_400_BAD_REQUEST)
-        content_type = (getattr(upload, 'content_type', '') or '').lower()
-        if content_type not in ALLOWED_MEDIA_UPLOAD_CONTENT_TYPES:
-            return Response({'file': ['Unsupported file type. Only image uploads are allowed.']}, status=status.HTTP_400_BAD_REQUEST)
-
-        rel_path = culture_media_upload_path(None, upload.name)
-        saved_path = default_storage.save(rel_path, upload)
-        media = MediaFile.objects.create(storage_path=saved_path)
-        record_entity_revision(
-            project=active_project,
-            entity_type='media_file',
-            object_id=media.pk,
-            action=EntityRevision.ACTION_CREATED,
-            snapshot=_serialize_instance(media),
-            user_name=_current_actor_label(request),
-        )
-        return Response({'id': media.id, 'storage_path': media.storage_path, 'uploaded_at': media.uploaded_at}, status=status.HTTP_201_CREATED)
-
-class NoteAttachmentListCreateView(APIView):
-    """List and upload image attachments for a planting plan note."""
-
-    parser_classes = [parsers.MultiPartParser, parsers.FormParser]
-
-    def get(self, request, note_id: int):
-        active_project = get_active_project_or_400(request)
-        plan = get_object_or_404(PlantingPlan, pk=note_id, project=active_project)
-        attachments = plan.attachments.all()
-        serializer = NoteAttachmentSerializer(attachments, many=True, context={'request': request})
-        return Response(serializer.data)
-
-    def post(self, request, note_id: int):
-        active_project = get_active_project_or_400(request)
-        plan = get_object_or_404(PlantingPlan, pk=note_id, project=active_project)
-
-        if plan.attachments.count() >= 10:
-            return Response({'detail': 'Attachment limit per note reached (10).'}, status=status.HTTP_400_BAD_REQUEST)
-
-        upload = request.FILES.get('image')
-        if upload is None:
-            return Response({'image': ['This field is required.']}, status=status.HTTP_400_BAD_REQUEST)
-
-        caption = request.data.get('caption', '')
-
-        try:
-            content, metadata = process_note_image(upload)
-        except ImageProcessingBackendUnavailableError as exc:
-            return Response({'detail': str(exc)}, status=status.HTTP_503_SERVICE_UNAVAILABLE)
-        except ImageProcessingError as exc:
-            return Response({'detail': str(exc)}, status=status.HTTP_400_BAD_REQUEST)
-
-        attachment = NoteAttachment(
-            planting_plan=plan,
-            caption=caption,
-            created_by=request.user if request.user.is_authenticated else None,
-            updated_by=request.user if request.user.is_authenticated else None,
-            project=plan.project,
-            width=metadata['width'],
-            height=metadata['height'],
-            size_bytes=metadata['size_bytes'],
-            mime_type=metadata['mime_type'],
-        )
-        attachment.image.save(str(metadata.get('filename', 'processed.webp')), content, save=False)
-        attachment.save()
-
-        serializer = NoteAttachmentSerializer(attachment, context={'request': request})
-        record_entity_revision(
-            project=attachment.project,
-            entity_type='note_attachment',
-            object_id=attachment.pk,
-            action=EntityRevision.ACTION_CREATED,
-            snapshot=_serialize_instance(attachment),
-            user_name=_current_actor_label(request),
-        )
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-
-class NoteAttachmentDeleteView(APIView):
-    """Delete a note attachment."""
-
-    def delete(self, request, attachment_id: int):
-        active_project = get_active_project_or_400(request)
-        attachment = get_object_or_404(NoteAttachment, pk=attachment_id, project=active_project)
-        attachment_id = attachment.pk
-        attachment_project = attachment.project
-        snapshot = _serialize_instance(attachment)
-        attachment.image.delete(save=False)
-        attachment.delete()
-        record_entity_revision(
-            project=attachment_project,
-            entity_type='note_attachment',
-            object_id=attachment_id,
-            action=EntityRevision.ACTION_DELETED,
-            snapshot=snapshot,
-            changed_fields=[],
-            user_name=_current_actor_label(request),
-        )
-        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class SeedDemandListView(ProjectScopedMixin, generics.ListAPIView):
